requirements.txt
pandas
selenium
numpy
unidecode
openpyxl
python-dotenv
